score usuarios

+++ database/migrations/2022_02_25_214922_create_employers_table.php
@@ -23,7 +23,7 @@ class CreateEmployersTable extends Migration
             $table->enum('sector', ['Restaurante', 'Bar','Comercio', 'Entretenimiento', 'Atención al cliente', 'Marketing','Tecnología','Otro'])->nullable(); // Sector? a que se dedica la empresa
             $table->string('description', 8000); // sirve en caso de que seleccione 'otro' en sector y para profundizar sobre que hace la empresa
             
-            $table->tinyInteger('score')->nullable();
+            $table->float('score')->nullable();
             $table->boolean('hidden')->default(false);
 
             $table->foreign('id_user')->references('id')->on('users')
